pyproject.toml
[project]
name = "zkteco-sync"
version = "0.1.0"
description = "ZKTeco device sync — ADMS listener + SDK poller + REST API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "pyzk>=0.9",
    "python-dotenv>=1.0.0",
]
